% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiome_barplot.R
\name{create_color_palette}
\alias{create_color_palette}
\title{create manual palette}
\usage{
create_color_palette(taxonomy_vec)
}
\arguments{
\item{taxonomy_vec}{}
}
\value{

}
\description{
create manual palette
}
