% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyankomicro-package.R
\docType{package}
\name{nyankomicro-package}
\alias{nyankomicro}
\alias{nyankomicro-package}
\title{nyankomicro: Useful Functions for Bacterial Composition Analysis}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Keigo Ide \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
