% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deseq2_tree.R
\name{deseq2_tree}
\alias{deseq2_tree}
\title{Show the results of deseq2 on the phylogenetic tree}
\usage{
deseq2_tree(
  physeq,
  deseq2_results,
  level = c("Kingdom", "Phylum", "Class", "Order", "Family", "Genus", "Species"),
  alpha = 0.05,
  sample_annotation = "Status",
  vis_domain = FALSE,
  colors = nyankocolors,
  tree_adjust = 1.2
)
}
\arguments{
\item{physeq}{phyloseq object}

\item{deseq2_results}{deseq2 results object}

\item{level}{visualize taxonomy level}

\item{alpha}{threshold of adjusted p-values}

\item{sample_annotation}{sample metadata used for drawing}

\item{vis_domain}{domain level}

\item{colors}{color palette}

\item{tree_adjust}{value for ggtree positioning.}
}
\value{
ggplot2 object
}
\description{
Show the results of deseq2 on the phylogenetic tree
}
